requirements.txt
Flask==2.2.5
Flask-SQLAlchemy==3.0.2
SQLAlchemy==1.4.41
python-dotenv==0.19.0
PyMySQL==1.0.2
Werkzeug==2.2.3
Bootstrap-Flask==2.2.0
Flask-WTF==1.0.0
Flask-Mail==0.9.1
APScheduler==3.10.1